# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ZagelMemoryResponsePreferences = { [key: string]: unknown };

export interface ZagelMemoryResponse {
  id: number;
  userId: string;
  name?: string | null;
  job?: string | null;
  interests?: string[];
  preferences?: ZagelMemoryResponsePreferences;
  behaviorRules?: string[];
  mood?: string;
  personality?: string;
}

export type UpdateZagelMemoryBodyPreferences = { [key: string]: unknown };

export interface UpdateZagelMemoryBody {
  userId?: string;
  name?: string;
  job?: string;
  interests?: string[];
  preferences?: UpdateZagelMemoryBodyPreferences;
  behaviorRules?: string[];
  mood?: string;
  personality?: string;
}

/**
 * User memory context
 */
export type ZagelChatBodyMemory = { [key: string]: unknown } | null;

export interface ZagelChatBody {
  message: string;
  userId?: string;
  context?: string;
  /** User memory context */
  memory?: ZagelChatBodyMemory;
}

export interface ZagelClassifyBody {
  text: string;
}

export type ZagelClassifyResponseType =
  (typeof ZagelClassifyResponseType)[keyof typeof ZagelClassifyResponseType];

export const ZagelClassifyResponseType = {
  instruction: "instruction",
  chat: "chat",
  onboarding: "onboarding",
  scenario: "scenario",
  template: "template",
} as const;

export interface ZagelClassifyResponse {
  type: ZagelClassifyResponseType;
  confidence: number;
  action?: string | null;
}

export interface Conversation {
  id: number;
  title?: string | null;
  createdAt: string;
}

export interface CreateConversationBody {
  title?: string | null;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface SendGeminiMessageBody {
  message: string;
  systemPrompt?: string | null;
}

export type GetZagelMemoryParams = {
  userId?: string;
};