# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: zagel
    description: Zagel AI assistant operations
  - name: gemini
    description: Gemini AI integration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /zagel/memory:
    get:
      operationId: getZagelMemory
      tags: [zagel]
      summary: Get user memory
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            default: "default"
      responses:
        "200":
          description: User memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZagelMemoryResponse"
    put:
      operationId: updateZagelMemory
      tags: [zagel]
      summary: Update user memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateZagelMemoryBody"
      responses:
        "200":
          description: Updated memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZagelMemoryResponse"

  /zagel/chat:
    post:
      operationId: zagelChat
      tags: [zagel]
      summary: Send message to Zagel AI (streaming SSE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZagelChatBody"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

  /zagel/classify:
    post:
      operationId: zagelClassify
      tags: [zagel]
      summary: Classify user input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZagelClassifyBody"
      responses:
        "200":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZagelClassifyResponse"

  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /gemini/conversations/{id}/messages:
    get:
      operationId: getGeminiConversationMessages
      tags: [gemini]
      summary: Get messages in conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGeminiMessageBody"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ZagelMemoryResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
          nullable: true
        job:
          type: string
          nullable: true
        interests:
          type: array
          items:
            type: string
        preferences:
          type: object
        behaviorRules:
          type: array
          items:
            type: string
        mood:
          type: string
        personality:
          type: string
      required:
        - id
        - userId

    UpdateZagelMemoryBody:
      type: object
      properties:
        userId:
          type: string
          default: "default"
        name:
          type: string
        job:
          type: string
        interests:
          type: array
          items:
            type: string
        preferences:
          type: object
        behaviorRules:
          type: array
          items:
            type: string
        mood:
          type: string
        personality:
          type: string

    ZagelChatBody:
      type: object
      properties:
        message:
          type: string
        userId:
          type: string
          default: "default"
        context:
          type: string
          default: "chat"
        memory:
          type: object
          nullable: true
          description: User memory context
      required:
        - message

    ZagelClassifyBody:
      type: object
      properties:
        text:
          type: string
      required:
        - text

    ZagelClassifyResponse:
      type: object
      properties:
        type:
          type: string
          enum: [instruction, chat, onboarding, scenario, template]
        confidence:
          type: number
        action:
          type: string
          nullable: true
      required:
        - type
        - confidence

    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - createdAt

    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
          nullable: true

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    SendGeminiMessageBody:
      type: object
      properties:
        message:
          type: string
        systemPrompt:
          type: string
          nullable: true
      required:
        - message