# lib/api-zod/src/generated/types/zagelClassifyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ZagelClassifyResponseType } from "./zagelClassifyResponseType";

export interface ZagelClassifyResponse {
  type: ZagelClassifyResponseType;
  confidence: number;
  action?: string | null;
}