# lib/api-zod/src/generated/types/zagelMemoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ZagelMemoryResponsePreferences } from "./zagelMemoryResponsePreferences";

export interface ZagelMemoryResponse {
  id: number;
  userId: string;
  name?: string | null;
  job?: string | null;
  interests?: string[];
  preferences?: ZagelMemoryResponsePreferences;
  behaviorRules?: string[];
  mood?: string;
  personality?: string;
}